fix!: Make `display_list` private
