+++ src/display_list/mod.rs
@@ -38,7 +38,7 @@ use std::{cmp, fmt};
 use crate::renderer::{stylesheet::Stylesheet, Margin, Style};
 
 /// List of lines to be displayed.
-pub struct DisplayList<'a> {
+pub(crate) struct DisplayList<'a> {
     pub body: Vec<DisplayLine<'a>>,
     pub stylesheet: Stylesheet,
     pub anonymized_line_numbers: bool,
@@ -343,7 +343,6 @@ impl<'a> DisplayList<'a> {
                 let indent_char = match annotation_part {
                     DisplayAnnotationPart::Standalone => ' ',
                     DisplayAnnotationPart::LabelContinuation => ' ',
-                    DisplayAnnotationPart::Consequitive => ' ',
                     DisplayAnnotationPart::MultilineStart => '_',
                     DisplayAnnotationPart::MultilineEnd => '_',
                 };
@@ -358,7 +357,6 @@ impl<'a> DisplayList<'a> {
                 let color = self.get_annotation_style(annotation_type);
                 let indent_length = match annotation_part {
                     DisplayAnnotationPart::LabelContinuation => range.1,
-                    DisplayAnnotationPart::Consequitive => range.1,
                     _ => range.0,
                 };
 
@@ -626,8 +624,6 @@ pub enum DisplayAnnotationPart {
     Standalone,
     /// A continuation of a multi-line label of an annotation.
     LabelContinuation,
-    /// A consequitive annotation in case multiple annotations annotate a single line.
-    Consequitive,
     /// A line starting a multiline annotation.
     MultilineStart,
     /// A line ending a multiline annotation.
@@ -1227,4 +1223,897 @@ fn is_annotation_empty(annotation: &Annotation<'_>) -> bool {
         .label
         .iter()
         .all(|fragment| fragment.content.is_empty())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_format_title() {
+        let input = snippet::Snippet {
+            title: Some(snippet::Annotation {
+                id: Some("E0001"),
+                label: Some("This is a title"),
+                annotation_type: snippet::AnnotationType::Error,
+            }),
+            footer: vec![],
+            slices: vec![],
+        };
+        let output = DisplayList {
+            body: vec![DisplayLine::Raw(DisplayRawLine::Annotation {
+                annotation: Annotation {
+                    annotation_type: DisplayAnnotationType::Error,
+                    id: Some("E0001"),
+                    label: vec![DisplayTextFragment {
+                        content: "This is a title",
+                        style: DisplayTextStyle::Emphasis,
+                    }],
+                },
+                source_aligned: false,
+                continuation: false,
+            })],
+            stylesheet: Stylesheet::default(),
+            anonymized_line_numbers: false,
+            margin: None,
+        };
+        assert_eq!(DisplayList::from(input), output);
+    }
+
+    #[test]
+    fn test_format_slice() {
+        let line_1 = "This is line 1";
+        let line_2 = "This is line 2";
+        let source = [line_1, line_2].join("\n");
+        let input = snippet::Snippet {
+            title: None,
+            footer: vec![],
+            slices: vec![snippet::Slice {
+                source: &source,
+                line_start: 5402,
+                origin: None,
+                annotations: vec![],
+                fold: false,
+            }],
+        };
+        let output = DisplayList {
+            body: vec![
+                DisplayLine::Source {
+                    lineno: None,
+                    inline_marks: vec![],
+                    line: DisplaySourceLine::Empty,
+                },
+                DisplayLine::Source {
+                    lineno: Some(5402),
+                    inline_marks: vec![],
+                    line: DisplaySourceLine::Content {
+                        text: line_1,
+                        range: (0, line_1.len()),
+                    },
+                },
+                DisplayLine::Source {
+                    lineno: Some(5403),
+                    inline_marks: vec![],
+                    line: DisplaySourceLine::Content {
+                        range: (line_1.len() + 1, source.len()),
+                        text: line_2,
+                    },
+                },
+                DisplayLine::Source {
+                    lineno: None,
+                    inline_marks: vec![],
+                    line: DisplaySourceLine::Empty,
+                },
+            ],
+            stylesheet: Stylesheet::default(),
+            anonymized_line_numbers: false,
+            margin: None,
+        };
+        assert_eq!(DisplayList::from(input), output);
+    }
+
+    #[test]
+    fn test_format_slices_continuation() {
+        let src_0 = "This is slice 1";
+        let src_0_len = src_0.len();
+        let src_1 = "This is slice 2";
+        let src_1_len = src_1.len();
+        let input = snippet::Snippet {
+            title: None,
+            footer: vec![],
+            slices: vec![
+                snippet::Slice {
+                    source: src_0,
+                    line_start: 5402,
+                    origin: Some("file1.rs"),
+                    annotations: vec![],
+                    fold: false,
+                },
+                snippet::Slice {
+                    source: src_1,
+                    line_start: 2,
+                    origin: Some("file2.rs"),
+                    annotations: vec![],
+                    fold: false,
+                },
+            ],
+        };
+        let output = DisplayList {
+            body: vec![
+                DisplayLine::Raw(DisplayRawLine::Origin {
+                    path: "file1.rs",
+                    pos: None,
+                    header_type: DisplayHeaderType::Initial,
+                }),
+                DisplayLine::Source {
+                    lineno: None,
+                    inline_marks: vec![],
+                    line: DisplaySourceLine::Empty,
+                },
+                DisplayLine::Source {
+                    lineno: Some(5402),
+                    inline_marks: vec![],
+                    line: DisplaySourceLine::Content {
+                        text: src_0,
+                        range: (0, src_0_len),
+                    },
+                },
+                DisplayLine::Source {
+                    lineno: None,
+                    inline_marks: vec![],
+                    line: DisplaySourceLine::Empty,
+                },
+                DisplayLine::Raw(DisplayRawLine::Origin {
+                    path: "file2.rs",
+                    pos: None,
+                    header_type: DisplayHeaderType::Continuation,
+                }),
+                DisplayLine::Source {
+                    lineno: None,
+                    inline_marks: vec![],
+                    line: DisplaySourceLine::Empty,
+                },
+                DisplayLine::Source {
+                    lineno: Some(2),
+                    inline_marks: vec![],
+                    line: DisplaySourceLine::Content {
+                        text: src_1,
+                        range: (0, src_1_len),
+                    },
+                },
+                DisplayLine::Source {
+                    lineno: None,
+                    inline_marks: vec![],
+                    line: DisplaySourceLine::Empty,
+                },
+            ],
+            stylesheet: Stylesheet::default(),
+            anonymized_line_numbers: false,
+            margin: None,
+        };
+        assert_eq!(DisplayList::from(input), output);
+    }
+
+    #[test]
+    fn test_format_slice_annotation_standalone() {
+        let line_1 = "This is line 1";
+        let line_2 = "This is line 2";
+        let source = [line_1, line_2].join("\n");
+        // In line 2
+        let range = (22, 24);
+        let input = snippet::Snippet {
+            title: None,
+            footer: vec![],
+            slices: vec![snippet::Slice {
+                source: &source,
+                line_start: 5402,
+                origin: None,
+                annotations: vec![snippet::SourceAnnotation {
+                    range,
+                    label: "Test annotation",
+                    annotation_type: snippet::AnnotationType::Info,
+                }],
+                fold: false,
+            }],
+        };
+        let output = DisplayList {
+            body: vec![
+                DisplayLine::Source {
+                    lineno: None,
+                    inline_marks: vec![],
+                    line: DisplaySourceLine::Empty,
+                },
+                DisplayLine::Source {
+                    lineno: Some(5402),
+                    inline_marks: vec![],
+                    line: DisplaySourceLine::Content {
+                        range: (0, line_1.len()),
+                        text: line_1,
+                    },
+                },
+                DisplayLine::Source {
+                    lineno: Some(5403),
+                    inline_marks: vec![],
+                    line: DisplaySourceLine::Content {
+                        range: (line_1.len() + 1, source.len()),
+                        text: line_2,
+                    },
+                },
+                DisplayLine::Source {
+                    lineno: None,
+                    inline_marks: vec![],
+                    line: DisplaySourceLine::Annotation {
+                        annotation: Annotation {
+                            annotation_type: DisplayAnnotationType::Info,
+                            id: None,
+                            label: vec![DisplayTextFragment {
+                                content: "Test annotation",
+                                style: DisplayTextStyle::Regular,
+                            }],
+                        },
+                        range: (range.0 - (line_1.len() + 1), range.1 - (line_1.len() + 1)),
+                        annotation_type: DisplayAnnotationType::Info,
+                        annotation_part: DisplayAnnotationPart::Standalone,
+                    },
+                },
+                DisplayLine::Source {
+                    lineno: None,
+                    inline_marks: vec![],
+                    line: DisplaySourceLine::Empty,
+                },
+            ],
+            stylesheet: Stylesheet::default(),
+            anonymized_line_numbers: false,
+            margin: None,
+        };
+        assert_eq!(DisplayList::from(input), output);
+    }
+
+    #[test]
+    fn test_format_label() {
+        let input = snippet::Snippet {
+            title: None,
+            footer: vec![snippet::Annotation {
+                id: None,
+                label: Some("This __is__ a title"),
+                annotation_type: snippet::AnnotationType::Error,
+            }],
+            slices: vec![],
+        };
+        let output = DisplayList {
+            body: vec![DisplayLine::Raw(DisplayRawLine::Annotation {
+                annotation: Annotation {
+                    annotation_type: DisplayAnnotationType::Error,
+                    id: None,
+                    label: vec![DisplayTextFragment {
+                        content: "This __is__ a title",
+                        style: DisplayTextStyle::Regular,
+                    }],
+                },
+                source_aligned: true,
+                continuation: false,
+            })],
+            stylesheet: Stylesheet::default(),
+            anonymized_line_numbers: false,
+            margin: None,
+        };
+        assert_eq!(DisplayList::from(input), output);
+    }
+
+    #[test]
+    #[should_panic]
+    fn test_i26() {
+        let source = "short";
+        let label = "label";
+        let input = snippet::Snippet {
+            title: None,
+            footer: vec![],
+            slices: vec![snippet::Slice {
+                annotations: vec![snippet::SourceAnnotation {
+                    range: (0, source.len() + 1),
+                    label,
+                    annotation_type: snippet::AnnotationType::Error,
+                }],
+                source,
+                line_start: 0,
+                origin: None,
+                fold: false,
+            }],
+        };
+
+        let _ = DisplayList::from(input);
+    }
+
+    #[test]
+    fn test_i_29() {
+        let snippets = snippet::Snippet {
+            title: Some(snippet::Annotation {
+                id: None,
+                label: Some("oops"),
+                annotation_type: snippet::AnnotationType::Error,
+            }),
+            footer: vec![],
+            slices: vec![snippet::Slice {
+                source: "First line\r\nSecond oops line",
+                line_start: 1,
+                origin: Some("<current file>"),
+                annotations: vec![snippet::SourceAnnotation {
+                    range: (19, 23),
+                    label: "oops",
+                    annotation_type: snippet::AnnotationType::Error,
+                }],
+                fold: true,
+            }],
+        };
+
+        let expected = DisplayList {
+            body: vec![
+                DisplayLine::Raw(DisplayRawLine::Annotation {
+                    annotation: Annotation {
+                        annotation_type: DisplayAnnotationType::Error,
+                        id: None,
+                        label: vec![DisplayTextFragment {
+                            content: "oops",
+                            style: DisplayTextStyle::Emphasis,
+                        }],
+                    },
+                    source_aligned: false,
+                    continuation: false,
+                }),
+                DisplayLine::Raw(DisplayRawLine::Origin {
+                    path: "<current file>",
+                    pos: Some((2, 8)),
+                    header_type: DisplayHeaderType::Initial,
+                }),
+                DisplayLine::Source {
+                    lineno: None,
+                    inline_marks: vec![],
+                    line: DisplaySourceLine::Empty,
+                },
+                DisplayLine::Source {
+                    lineno: Some(1),
+                    inline_marks: vec![],
+                    line: DisplaySourceLine::Content {
+                        text: "First line",
+                        range: (0, 10),
+                    },
+                },
+                DisplayLine::Source {
+                    lineno: Some(2),
+                    inline_marks: vec![],
+                    line: DisplaySourceLine::Content {
+                        text: "Second oops line",
+                        range: (12, 28),
+                    },
+                },
+                DisplayLine::Source {
+                    lineno: None,
+                    inline_marks: vec![],
+                    line: DisplaySourceLine::Annotation {
+                        annotation: Annotation {
+                            annotation_type: DisplayAnnotationType::None,
+                            id: None,
+                            label: vec![DisplayTextFragment {
+                                content: "oops",
+                                style: DisplayTextStyle::Regular,
+                            }],
+                        },
+                        range: (7, 11),
+                        annotation_type: DisplayAnnotationType::Error,
+                        annotation_part: DisplayAnnotationPart::Standalone,
+                    },
+                },
+                DisplayLine::Source {
+                    lineno: None,
+                    inline_marks: vec![],
+                    line: DisplaySourceLine::Empty,
+                },
+            ],
+            stylesheet: Stylesheet::default(),
+            anonymized_line_numbers: false,
+            margin: None,
+        };
+
+        assert_eq!(DisplayList::from(snippets), expected);
+    }
+
+    #[test]
+    fn test_source_empty() {
+        let dl = DisplayList::from(vec![DisplayLine::Source {
+            lineno: None,
+            inline_marks: vec![],
+            line: DisplaySourceLine::Empty,
+        }]);
+
+        assert_eq!(dl.to_string(), " |");
+    }
+
+    #[test]
+    fn test_source_content() {
+        let dl = DisplayList::from(vec![
+            DisplayLine::Source {
+                lineno: Some(56),
+                inline_marks: vec![],
+                line: DisplaySourceLine::Content {
+                    text: "This is an example",
+                    range: (0, 19),
+                },
+            },
+            DisplayLine::Source {
+                lineno: Some(57),
+                inline_marks: vec![],
+                line: DisplaySourceLine::Content {
+                    text: "of content lines",
+                    range: (0, 19),
+                },
+            },
+        ]);
+
+        assert_eq!(
+            dl.to_string(),
+            "56 | This is an example\n57 | of content lines"
+        );
+    }
+
+    #[test]
+    fn test_source_annotation_standalone_singleline() {
+        let dl = DisplayList::from(vec![DisplayLine::Source {
+            lineno: None,
+            inline_marks: vec![],
+            line: DisplaySourceLine::Annotation {
+                range: (0, 5),
+                annotation: Annotation {
+                    annotation_type: DisplayAnnotationType::None,
+                    id: None,
+                    label: vec![DisplayTextFragment {
+                        content: "Example string",
+                        style: DisplayTextStyle::Regular,
+                    }],
+                },
+                annotation_type: DisplayAnnotationType::Error,
+                annotation_part: DisplayAnnotationPart::Standalone,
+            },
+        }]);
+
+        assert_eq!(dl.to_string(), " | ^^^^^ Example string");
+    }
+
+    #[test]
+    fn test_source_annotation_standalone_multiline() {
+        let dl = DisplayList::from(vec![
+            DisplayLine::Source {
+                lineno: None,
+                inline_marks: vec![],
+                line: DisplaySourceLine::Annotation {
+                    range: (0, 5),
+                    annotation: Annotation {
+                        annotation_type: DisplayAnnotationType::Help,
+                        id: None,
+                        label: vec![DisplayTextFragment {
+                            content: "Example string",
+                            style: DisplayTextStyle::Regular,
+                        }],
+                    },
+                    annotation_type: DisplayAnnotationType::Warning,
+                    annotation_part: DisplayAnnotationPart::Standalone,
+                },
+            },
+            DisplayLine::Source {
+                lineno: None,
+                inline_marks: vec![],
+                line: DisplaySourceLine::Annotation {
+                    range: (0, 5),
+                    annotation: Annotation {
+                        annotation_type: DisplayAnnotationType::Help,
+                        id: None,
+                        label: vec![DisplayTextFragment {
+                            content: "Second line",
+                            style: DisplayTextStyle::Regular,
+                        }],
+                    },
+                    annotation_type: DisplayAnnotationType::Warning,
+                    annotation_part: DisplayAnnotationPart::LabelContinuation,
+                },
+            },
+        ]);
+
+        assert_eq!(
+            dl.to_string(),
+            " | ----- help: Example string\n |             Second line"
+        );
+    }
+
+    #[test]
+    fn test_source_annotation_standalone_multi_annotation() {
+        let dl = DisplayList::from(vec![
+            DisplayLine::Source {
+                lineno: None,
+                inline_marks: vec![],
+                line: DisplaySourceLine::Annotation {
+                    range: (0, 5),
+                    annotation: Annotation {
+                        annotation_type: DisplayAnnotationType::Info,
+                        id: None,
+                        label: vec![DisplayTextFragment {
+                            content: "Example string",
+                            style: DisplayTextStyle::Regular,
+                        }],
+                    },
+                    annotation_type: DisplayAnnotationType::Note,
+                    annotation_part: DisplayAnnotationPart::Standalone,
+                },
+            },
+            DisplayLine::Source {
+                lineno: None,
+                inline_marks: vec![],
+                line: DisplaySourceLine::Annotation {
+                    range: (0, 5),
+                    annotation: Annotation {
+                        annotation_type: DisplayAnnotationType::Info,
+                        id: None,
+                        label: vec![DisplayTextFragment {
+                            content: "Second line",
+                            style: DisplayTextStyle::Regular,
+                        }],
+                    },
+                    annotation_type: DisplayAnnotationType::Note,
+                    annotation_part: DisplayAnnotationPart::LabelContinuation,
+                },
+            },
+            DisplayLine::Source {
+                lineno: None,
+                inline_marks: vec![],
+                line: DisplaySourceLine::Annotation {
+                    range: (0, 5),
+                    annotation: Annotation {
+                        annotation_type: DisplayAnnotationType::Warning,
+                        id: None,
+                        label: vec![DisplayTextFragment {
+                            content: "Second line of the warning",
+                            style: DisplayTextStyle::Regular,
+                        }],
+                    },
+                    annotation_type: DisplayAnnotationType::Note,
+                    annotation_part: DisplayAnnotationPart::LabelContinuation,
+                },
+            },
+            DisplayLine::Source {
+                lineno: None,
+                inline_marks: vec![],
+                line: DisplaySourceLine::Annotation {
+                    range: (0, 5),
+                    annotation: Annotation {
+                        annotation_type: DisplayAnnotationType::Info,
+                        id: None,
+                        label: vec![DisplayTextFragment {
+                            content: "This is an info",
+                            style: DisplayTextStyle::Regular,
+                        }],
+                    },
+                    annotation_type: DisplayAnnotationType::Info,
+                    annotation_part: DisplayAnnotationPart::Standalone,
+                },
+            },
+            DisplayLine::Source {
+                lineno: None,
+                inline_marks: vec![],
+                line: DisplaySourceLine::Annotation {
+                    range: (0, 5),
+                    annotation: Annotation {
+                        annotation_type: DisplayAnnotationType::Help,
+                        id: None,
+                        label: vec![DisplayTextFragment {
+                            content: "This is help",
+                            style: DisplayTextStyle::Regular,
+                        }],
+                    },
+                    annotation_type: DisplayAnnotationType::Help,
+                    annotation_part: DisplayAnnotationPart::Standalone,
+                },
+            },
+            DisplayLine::Source {
+                lineno: None,
+                inline_marks: vec![],
+                line: DisplaySourceLine::Annotation {
+                    range: (0, 0),
+                    annotation: Annotation {
+                        annotation_type: DisplayAnnotationType::None,
+                        id: None,
+                        label: vec![DisplayTextFragment {
+                            content: "This is an annotation of type none",
+                            style: DisplayTextStyle::Regular,
+                        }],
+                    },
+                    annotation_type: DisplayAnnotationType::None,
+                    annotation_part: DisplayAnnotationPart::Standalone,
+                },
+            },
+        ]);
+
+        assert_eq!(dl.to_string(), " | ----- info: Example string\n |             Second line\n |                Second line of the warning\n | ----- info: This is an info\n | ----- help: This is help\n |  This is an annotation of type none");
+    }
+
+    #[test]
+    fn test_fold_line() {
+        let dl = DisplayList::from(vec![
+            DisplayLine::Source {
+                lineno: Some(5),
+                inline_marks: vec![],
+                line: DisplaySourceLine::Content {
+                    text: "This is line 5",
+                    range: (0, 19),
+                },
+            },
+            DisplayLine::Fold {
+                inline_marks: vec![],
+            },
+            DisplayLine::Source {
+                lineno: Some(10021),
+                inline_marks: vec![],
+                line: DisplaySourceLine::Content {
+                    text: "... and now we're at line 10021",
+                    range: (0, 19),
+                },
+            },
+        ]);
+
+        assert_eq!(
+            dl.to_string(),
+            "    5 | This is line 5\n...\n10021 | ... and now we're at line 10021"
+        );
+    }
+
+    #[test]
+    fn test_raw_origin_initial_nopos() {
+        let dl = DisplayList::from(vec![DisplayLine::Raw(DisplayRawLine::Origin {
+            path: "src/test.rs",
+            pos: None,
+            header_type: DisplayHeaderType::Initial,
+        })]);
+
+        assert_eq!(dl.to_string(), "--> src/test.rs");
+    }
+
+    #[test]
+    fn test_raw_origin_initial_pos() {
+        let dl = DisplayList::from(vec![DisplayLine::Raw(DisplayRawLine::Origin {
+            path: "src/test.rs",
+            pos: Some((23, 15)),
+            header_type: DisplayHeaderType::Initial,
+        })]);
+
+        assert_eq!(dl.to_string(), "--> src/test.rs:23:15");
+    }
+
+    #[test]
+    fn test_raw_origin_continuation() {
+        let dl = DisplayList::from(vec![DisplayLine::Raw(DisplayRawLine::Origin {
+            path: "src/test.rs",
+            pos: Some((23, 15)),
+            header_type: DisplayHeaderType::Continuation,
+        })]);
+
+        assert_eq!(dl.to_string(), "::: src/test.rs:23:15");
+    }
+
+    #[test]
+    fn test_raw_annotation_unaligned() {
+        let dl = DisplayList::from(vec![DisplayLine::Raw(DisplayRawLine::Annotation {
+            annotation: Annotation {
+                annotation_type: DisplayAnnotationType::Error,
+                id: Some("E0001"),
+                label: vec![DisplayTextFragment {
+                    content: "This is an error",
+                    style: DisplayTextStyle::Regular,
+                }],
+            },
+            source_aligned: false,
+            continuation: false,
+        })]);
+
+        assert_eq!(dl.to_string(), "error[E0001]: This is an error");
+    }
+
+    #[test]
+    fn test_raw_annotation_unaligned_multiline() {
+        let dl = DisplayList::from(vec![
+            DisplayLine::Raw(DisplayRawLine::Annotation {
+                annotation: Annotation {
+                    annotation_type: DisplayAnnotationType::Warning,
+                    id: Some("E0001"),
+                    label: vec![DisplayTextFragment {
+                        content: "This is an error",
+                        style: DisplayTextStyle::Regular,
+                    }],
+                },
+                source_aligned: false,
+                continuation: false,
+            }),
+            DisplayLine::Raw(DisplayRawLine::Annotation {
+                annotation: Annotation {
+                    annotation_type: DisplayAnnotationType::Warning,
+                    id: Some("E0001"),
+                    label: vec![DisplayTextFragment {
+                        content: "Second line of the error",
+                        style: DisplayTextStyle::Regular,
+                    }],
+                },
+                source_aligned: false,
+                continuation: true,
+            }),
+        ]);
+
+        assert_eq!(
+            dl.to_string(),
+            "warning[E0001]: This is an error\n                Second line of the error"
+        );
+    }
+
+    #[test]
+    fn test_raw_annotation_aligned() {
+        let dl = DisplayList::from(vec![DisplayLine::Raw(DisplayRawLine::Annotation {
+            annotation: Annotation {
+                annotation_type: DisplayAnnotationType::Error,
+                id: Some("E0001"),
+                label: vec![DisplayTextFragment {
+                    content: "This is an error",
+                    style: DisplayTextStyle::Regular,
+                }],
+            },
+            source_aligned: true,
+            continuation: false,
+        })]);
+
+        assert_eq!(dl.to_string(), " = error[E0001]: This is an error");
+    }
+
+    #[test]
+    fn test_raw_annotation_aligned_multiline() {
+        let dl = DisplayList::from(vec![
+            DisplayLine::Raw(DisplayRawLine::Annotation {
+                annotation: Annotation {
+                    annotation_type: DisplayAnnotationType::Warning,
+                    id: Some("E0001"),
+                    label: vec![DisplayTextFragment {
+                        content: "This is an error",
+                        style: DisplayTextStyle::Regular,
+                    }],
+                },
+                source_aligned: true,
+                continuation: false,
+            }),
+            DisplayLine::Raw(DisplayRawLine::Annotation {
+                annotation: Annotation {
+                    annotation_type: DisplayAnnotationType::Warning,
+                    id: Some("E0001"),
+                    label: vec![DisplayTextFragment {
+                        content: "Second line of the error",
+                        style: DisplayTextStyle::Regular,
+                    }],
+                },
+                source_aligned: true,
+                continuation: true,
+            }),
+        ]);
+
+        assert_eq!(
+            dl.to_string(),
+            " = warning[E0001]: This is an error\n                   Second line of the error"
+        );
+    }
+
+    #[test]
+    fn test_different_annotation_types() {
+        let dl = DisplayList::from(vec![
+            DisplayLine::Raw(DisplayRawLine::Annotation {
+                annotation: Annotation {
+                    annotation_type: DisplayAnnotationType::Note,
+                    id: None,
+                    label: vec![DisplayTextFragment {
+                        content: "This is a note",
+                        style: DisplayTextStyle::Regular,
+                    }],
+                },
+                source_aligned: false,
+                continuation: false,
+            }),
+            DisplayLine::Raw(DisplayRawLine::Annotation {
+                annotation: Annotation {
+                    annotation_type: DisplayAnnotationType::None,
+                    id: None,
+                    label: vec![DisplayTextFragment {
+                        content: "This is just a string",
+                        style: DisplayTextStyle::Regular,
+                    }],
+                },
+                source_aligned: false,
+                continuation: false,
+            }),
+            DisplayLine::Raw(DisplayRawLine::Annotation {
+                annotation: Annotation {
+                    annotation_type: DisplayAnnotationType::None,
+                    id: None,
+                    label: vec![DisplayTextFragment {
+                        content: "Second line of none type annotation",
+                        style: DisplayTextStyle::Regular,
+                    }],
+                },
+                source_aligned: false,
+                continuation: true,
+            }),
+        ]);
+
+        assert_eq!(
+            dl.to_string(),
+            "note: This is a note\nThis is just a string\n  Second line of none type annotation",
+        );
+    }
+
+    #[test]
+    fn test_inline_marks_empty_line() {
+        let dl = DisplayList::from(vec![DisplayLine::Source {
+            lineno: None,
+            inline_marks: vec![DisplayMark {
+                mark_type: DisplayMarkType::AnnotationThrough,
+                annotation_type: DisplayAnnotationType::Error,
+            }],
+            line: DisplaySourceLine::Empty,
+        }]);
+
+        assert_eq!(dl.to_string(), " | |",);
+    }
+
+    #[test]
+    fn test_anon_lines() {
+        let mut dl = DisplayList::from(vec![
+            DisplayLine::Source {
+                lineno: Some(56),
+                inline_marks: vec![],
+                line: DisplaySourceLine::Content {
+                    text: "This is an example",
+                    range: (0, 19),
+                },
+            },
+            DisplayLine::Source {
+                lineno: Some(57),
+                inline_marks: vec![],
+                line: DisplaySourceLine::Content {
+                    text: "of content lines",
+                    range: (0, 19),
+                },
+            },
+            DisplayLine::Source {
+                lineno: None,
+                inline_marks: vec![],
+                line: DisplaySourceLine::Empty,
+            },
+            DisplayLine::Source {
+                lineno: None,
+                inline_marks: vec![],
+                line: DisplaySourceLine::Content {
+                    text: "abc",
+                    range: (0, 19),
+                },
+            },
+        ]);
+
+        dl.anonymized_line_numbers = true;
+        assert_eq!(
+            dl.to_string(),
+            "LL | This is an example\nLL | of content lines\n   |\n   | abc"
+        );
+    }
+
+    #[test]
+    fn test_raw_origin_initial_pos_anon_lines() {
+        let mut dl = DisplayList::from(vec![DisplayLine::Raw(DisplayRawLine::Origin {
+            path: "src/test.rs",
+            pos: Some((23, 15)),
+            header_type: DisplayHeaderType::Initial,
+        })]);
+
+        // Using anonymized_line_numbers should not affect the initial position
+        dl.anonymized_line_numbers = true;
+        assert_eq!(dl.to_string(), "--> src/test.rs:23:15");
+    }
 }
